+++ CesiumGltf/test/TestPropertyTexturePropertyView.cpp
@@ -1771,8 +1771,8 @@ TEST_CASE("Check sampling with different sampler values") {
   }
 }
 
-TEST_CASE(
-    "Test PropertyTextureProperty constructs with KHR_texture_transform") {
+TEST_CASE("Test PropertyTextureProperty constructs with "
+          "applyKhrTextureTransformExtension = true") {
   std::vector<uint8_t> data{1, 2, 3, 4};
 
   PropertyTextureProperty property;
@@ -1780,9 +1780,9 @@ TEST_CASE(
 
   ExtensionKhrTextureTransform& textureTransformExtension =
       property.addExtension<ExtensionKhrTextureTransform>();
-  textureTransformExtension.offset = {1.0, 2.0};
+  textureTransformExtension.offset = {0.5, -0.5};
   textureTransformExtension.rotation = CesiumUtility::Math::PiOverTwo;
-  textureTransformExtension.scale = {2.0, 0.5};
+  textureTransformExtension.scale = {0.5, 0.5};
   textureTransformExtension.texCoord = 10;
 
   ClassProperty classProperty;
@@ -1805,27 +1805,43 @@ TEST_CASE(
 
   property.channels = {0};
 
-  PropertyTexturePropertyView<uint8_t> view(
-      property,
-      classProperty,
-      sampler,
-      image);
+  PropertyTexturePropertyView<uint8_t>
+      view(property, classProperty, sampler, image, true);
   REQUIRE(view.status() == PropertyTexturePropertyViewStatus::Valid);
 
   auto textureTransform = view.getTextureTransform();
   REQUIRE(textureTransform != std::nullopt);
-  REQUIRE(textureTransform->offset() == glm::dvec2(1.0, 2.0));
+  REQUIRE(textureTransform->offset() == glm::dvec2(0.5, -0.5));
   REQUIRE(textureTransform->rotation() == CesiumUtility::Math::PiOverTwo);
-  REQUIRE(textureTransform->scale() == glm::dvec2(2.0, 0.5));
+  REQUIRE(textureTransform->scale() == glm::dvec2(0.5, 0.5));
 
-  // Texcoord is not overridden by value in KHR_texture_transform, in case that
-  // clients don't support the extension.
-  REQUIRE(view.getTexCoordSetIndex() == 0);
+  // Texcoord is overridden by value in KHR_texture_transform.
+  REQUIRE(
+      view.getTexCoordSetIndex() == textureTransform->getTexCoordSetIndex());
   REQUIRE(textureTransform->getTexCoordSetIndex() == 10);
+
+  // This transforms to the following UV values:
+  // (0, 0) -> (0.5, -0.5) -> wraps to (0.5, 0.5)
+  // (1, 0) -> (0.5, -1) -> wraps to (0.5, 0)
+  // (0, 1) -> (1, -0.5) -> wraps to (0, 0.5)
+  // (1, 1) -> (1, -1) -> wraps to (0.0, 0.0)
+  std::vector<glm::dvec2> texCoords{
+      glm::dvec2(0, 0),
+      glm::dvec2(1.0, 0),
+      glm::dvec2(0, 1.0),
+      glm::dvec2(1.0, 1.0)};
+
+  std::vector<uint8_t> expectedValues{4, 2, 3, 1};
+
+  for (size_t i = 0; i < texCoords.size(); i++) {
+    glm::dvec2 uv = texCoords[i];
+    REQUIRE(view.getRaw(uv[0], uv[1]) == expectedValues[i]);
+    REQUIRE(view.get(uv[0], uv[1]) == expectedValues[i]);
+  }
 }
 
 TEST_CASE("Test normalized PropertyTextureProperty constructs with "
-          "KHR_texture_transform") {
+          "applyKhrTextureTransformExtension = true") {
   std::vector<uint8_t> data{0, 64, 127, 255};
 
   PropertyTextureProperty property;
@@ -1833,9 +1849,9 @@ TEST_CASE("Test normalized PropertyTextureProperty constructs with "
 
   ExtensionKhrTextureTransform& textureTransformExtension =
       property.addExtension<ExtensionKhrTextureTransform>();
-  textureTransformExtension.offset = {1.0, 2.0};
+  textureTransformExtension.offset = {0.5, -0.5};
   textureTransformExtension.rotation = CesiumUtility::Math::PiOverTwo;
-  textureTransformExtension.scale = {2.0, 0.5};
+  textureTransformExtension.scale = {0.5, 0.5};
   textureTransformExtension.texCoord = 10;
 
   ClassProperty classProperty;
@@ -1859,21 +1875,39 @@ TEST_CASE("Test normalized PropertyTextureProperty constructs with "
 
   property.channels = {0};
 
-  PropertyTexturePropertyView<uint8_t, true> view(
-      property,
-      classProperty,
-      sampler,
-      image);
+  PropertyTexturePropertyView<uint8_t, true>
+      view(property, classProperty, sampler, image, true);
   REQUIRE(view.status() == PropertyTexturePropertyViewStatus::Valid);
 
   auto textureTransform = view.getTextureTransform();
   REQUIRE(textureTransform != std::nullopt);
-  REQUIRE(textureTransform->offset() == glm::dvec2(1.0, 2.0));
+  REQUIRE(textureTransform->offset() == glm::dvec2(0.5, -0.5));
   REQUIRE(textureTransform->rotation() == CesiumUtility::Math::PiOverTwo);
-  REQUIRE(textureTransform->scale() == glm::dvec2(2.0, 0.5));
+  REQUIRE(textureTransform->scale() == glm::dvec2(0.5, 0.5));
 
-  // Texcoord is not overridden by value in KHR_texture_transform, in case that
-  // clients don't support the extension.
-  REQUIRE(view.getTexCoordSetIndex() == 0);
+  // Texcoord is overridden by value in KHR_texture_transform.
+  REQUIRE(
+      view.getTexCoordSetIndex() == textureTransform->getTexCoordSetIndex());
   REQUIRE(textureTransform->getTexCoordSetIndex() == 10);
+
+  // This transforms to the following UV values:
+  // (0, 0) -> (0.5, -0.5) -> wraps to (0.5, 0.5)
+  // (1, 0) -> (0.5, -1) -> wraps to (0.5, 0)
+  // (0, 1) -> (1, -0.5) -> wraps to (0, 0.5)
+  // (1, 1) -> (1, -1) -> wraps to (0.0, 0.0)
+  std::vector<glm::dvec2> texCoords{
+      glm::dvec2(0, 0),
+      glm::dvec2(1.0, 0),
+      glm::dvec2(0, 1.0),
+      glm::dvec2(1.0, 1.0)};
+
+  std::vector<uint8_t> expectedValues{255, 64, 127, 0};
+
+  for (size_t i = 0; i < texCoords.size(); i++) {
+    glm::dvec2 uv = texCoords[i];
+    REQUIRE(view.getRaw(uv[0], uv[1]) == expectedValues[i]);
+    REQUIRE(
+        view.get(uv[0], uv[1]) ==
+        static_cast<double>(expectedValues[i]) / 255.0);
+  }
 }